requirements.txt
anyio==3.6.2
black==23.3.0
certifi==2022.12.7
click==8.1.3
contourpy==1.1.0
cycler==0.11.0
dnspython==2.3.0
email-validator==2.0.0.post2
fastapi==0.95.1
fonttools==4.40.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
mypy-extensions==1.0.0
numpy==1.24.3
orjson==3.8.11
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.0
pydantic==1.10.7
PyJWT==2.7.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.18
starlette==0.26.1
typing_extensions==4.5.0
ujson==5.7.0
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.2
